Changed traverse function to be friendly towards filtered lists and changed up arrow key functionality
The traverse function now works for filtered lists. (It only looks for elements that are currently visible while traversing. 

I have also changed the up/left arrow keys to move up one at a time instead of going to the top of the list.

You can also now press the 'f' key (or ctrl-f) while traversing the list to move focus back to the filter.

You will also notice an id attached to the button (line 56) that helps to allow auto-focus on the filter when you open the widget.

You'll also need to add this option on initialization of the widget:

open: function() {			$(this).multiselect("widget").find("input[type='search']:first").focus();
}